requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

python-dotenv==1.0.0
pydantic[email]==2.5.3

sqlalchemy==2.0.25
psycopg2-binary==2.9.9

passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0

boto3==1.34.18
botocore==1.34.18

numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
timm==0.9.12

opencv-python-headless==4.8.1.78

# pytorch-grad-cam==1.4.8